The static version works...
