+++ src/bokeh_joystick_widget/.bokeh
@@ -1,1 +1,1 @@
-{"bokeh_version":"3.4.1","signatures":{"package.json":"1740dadf52510c3adb9f83aa668a5870ad1a7e941cc51dae3f039874af7b3894","package-lock.json":"75e1f39c793ef76c00b5bdcc2d180e4de2613525c5a5807e9aa7536955bd497d","tsconfig.json":"64cb4b36963ed807db12a16eaec8b1796792178996e0f0044c39c82bbd9403c1"}}
+{"bokeh_version":"3.4.1","signatures":{"package.json":"343619451d37d8b5986bd105e9b8080779be0551a89553f3d05449b78ff5e60c","package-lock.json":"03ab0f0153f035d54184607bfceb1fcbbddc061530f5997bd33d158b1d1c566d","tsconfig.json":"ee03023ea9d772b7d337b63c15d95429817bc02fffd66ae17cd3c34d5f705185"}}
