+++ src/bokeh_joystick_widget/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "bokeh_joystick",
+  "name": "bokeh_joystick_widget",
   "version": "0.0.1",
   "description": "",
   "license": "BSD-3-Clause",
@@ -8,7 +8,5 @@
   "dependencies": {
     "@bokeh/bokehjs": "^3.4.1"
   },
-  "devDependencies": {
-    "typescript": "~5.4.3"
-  }
+  "devDependencies": {}
 }
