+++ src/bokeh_joystick_widget/tsconfig.json
@@ -5,7 +5,7 @@
     "noImplicitReturns": true,
     "noUnusedLocals": true,
     "noUnusedParameters": true,
-    "noImplicitOverride": true,
+    "noImplicitOverride": false,
     "strictNullChecks": true,
     "strictBindCallApply": true,
     "strictFunctionTypes": true,
@@ -23,11 +23,16 @@
     "module": "ES2022",
     "moduleResolution": "node",
     "isolatedModules": true,
+    "verbatimModuleSyntax": false,
     "resolveJsonModule": true,
-    "esModuleInterop": true,
     "skipLibCheck": true,
+    "esModuleInterop": true,
     "target": "ES2022",
-    "lib": ["es2022", "dom", "dom.iterable"],
+    "lib": [
+      "es2022",
+      "dom",
+      "dom.iterable"
+    ],
     "baseUrl": ".",
     "outDir": "./dist/lib",
     "paths": {
@@ -36,5 +41,7 @@
       ]
     }
   },
-  "include": ["./**/*.ts"]
+  "include": [
+    "./**/*.ts"
+  ],
 }
